move user funcs to users.go
