+++ server/auth.go
@@ -76,86 +76,6 @@ func login(pRuntime *runtime.Runtime) gin.HandlerFunc {
 	}
 }
 
-func updateUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
-	return func(c *gin.Context) {
-
-		if auth := c.GetBool("authenticated"); !auth {
-			c.JSON(http.StatusUnauthorized, gin.H{"error": "authorization required"})
-			return
-		}
-
-		up := &authUserUpdateInput{}
-
-		if err := c.ShouldBindJSON(up); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-			return
-		}
-
-		//------------------
-		// UPDATE
-		gErr := userUpdateDb(up, c, pRuntime)
-		if gErr != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
-			return
-		}
-		//------------------
-		// OUTPUT
-		c.Status(http.StatusOK)
-	}
-}
-
-func getUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
-	return func(c *gin.Context) {
-
-		auth := c.GetBool("authenticated")
-
-		userIdStr := c.Query("user_id")
-		input := &authUserGetInput{
-			UserId: persist.DbId(userIdStr),
-		}
-
-		output, gErr := userGetDb(input,
-			auth,
-			c, pRuntime)
-		if gErr != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
-			return
-		}
-
-		//------------------
-		// OUTPUT
-
-		c.JSON(http.StatusOK, output)
-
-	}
-}
-
-func createUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
-	return func(c *gin.Context) {
-
-		input := &authUserCreateInput{}
-
-		if err := c.ShouldBindJSON(input); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-			return
-		}
-
-		//------------------
-		// USER_CREATE
-		output, gErr := userCreateDb(input, c, pRuntime)
-		if gErr != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
-			return
-		}
-
-		//------------------
-		// OUTPUT
-
-		c.JSON(http.StatusOK, output)
-
-	}
-}
-
 //-------------------------------------------------------------
 // NONCE
 //-------------------------------------------------------------
