+++ server/user.go
@@ -2,7 +2,9 @@ package server
 
 import (
 	"context"
+	"net/http"
 
+	"github.com/gin-gonic/gin"
 	"github.com/mikeydub/go-gallery/persist"
 	"github.com/mikeydub/go-gallery/runtime"
 	// "github.com/davecgh/go-spew/spew"
@@ -10,19 +12,19 @@ import (
 
 //-------------------------------------------------------------
 // INPUT - USER_UPDATE
-type authUserUpdateInput struct {
+type userUpdateInput struct {
 	UserId      persist.DbId `json:"address" validate:"required,eth_addr"` // len=42"` // standard ETH "0x"-prefixed address
 	UserNameStr string       `json:"username"`
 	BioStr      string       `json:"description"`
 }
 
 // INPUT - USER_GET
-type authUserGetInput struct {
+type userGetInput struct {
 	UserId persist.DbId `json:"user_id" validate:"required,eth_addr"` // len=42"` // standard ETH "0x"-prefixed address
 }
 
 // OUTPUT - USER_GET
-type authUserGetOutput struct {
+type userGetOutput struct {
 	UserNameStr string ` json:"username"`
 	BioStr      string ` json:"bio"`
 	Address     string ` json:"address"`
@@ -57,7 +59,7 @@ type authUserGetPreflightOutput struct {
 //         this is to allow for users interupting the onboarding flow, and to be able to come back to it later
 //         and the system recognize that their user already exists.
 //         the users entering details on the user as they onboard are all user-update operations.
-type authUserCreateInput struct {
+type userCreateInput struct {
 
 	// needed because this is a new user that cant be logged into, and the client creating
 	// the user still needs to prove ownership of their address.
@@ -67,17 +69,100 @@ type authUserCreateInput struct {
 }
 
 // OUTPUT - USER_CREATE
-type authUserCreateOutput struct {
+type userCreateOutput struct {
 	SignatureValidBool bool         `json:"signature_valid"`
 	JWTtokenStr        string       `json:"jwt_token"` // JWT token is sent back to user to use to continue onboarding
 	UserIDstr          persist.DbId `json:"user_id"`
 }
 
 //-------------------------------------------------------------
+// HANDLERS
+
+func updateUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
+	return func(c *gin.Context) {
+
+		if auth := c.GetBool("authenticated"); !auth {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "authorization required"})
+			return
+		}
+
+		up := &userUpdateInput{}
+
+		if err := c.ShouldBindJSON(up); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		//------------------
+		// UPDATE
+		gErr := userUpdateDb(up, c, pRuntime)
+		if gErr != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
+			return
+		}
+		//------------------
+		// OUTPUT
+		c.Status(http.StatusOK)
+	}
+}
+
+func getUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
+	return func(c *gin.Context) {
+
+		auth := c.GetBool("authenticated")
+
+		userIdStr := c.Query("user_id")
+		input := &userGetInput{
+			UserId: persist.DbId(userIdStr),
+		}
+
+		output, gErr := userGetDb(input,
+			auth,
+			c, pRuntime)
+		if gErr != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
+			return
+		}
+
+		//------------------
+		// OUTPUT
+
+		c.JSON(http.StatusOK, output)
+
+	}
+}
+
+func createUserAuth(pRuntime *runtime.Runtime) gin.HandlerFunc {
+	return func(c *gin.Context) {
+
+		input := &userCreateInput{}
+
+		if err := c.ShouldBindJSON(input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		//------------------
+		// USER_CREATE
+		output, gErr := userCreateDb(input, c, pRuntime)
+		if gErr != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": gErr})
+			return
+		}
+
+		//------------------
+		// OUTPUT
+
+		c.JSON(http.StatusOK, output)
+
+	}
+}
+
+//-------------------------------------------------------------
 // USER_CREATE__PIPELINE
-func userCreateDb(pInput *authUserCreateInput,
+func userCreateDb(pInput *userCreateInput,
 	pCtx context.Context,
-	pRuntime *runtime.Runtime) (*authUserCreateOutput, error) {
+	pRuntime *runtime.Runtime) (*userCreateOutput, error) {
 
 	//------------------
 	// VALIDATE
@@ -87,7 +172,7 @@ func userCreateDb(pInput *authUserCreateInput,
 	}
 
 	//------------------
-	output := &authUserCreateOutput{}
+	output := &userCreateOutput{}
 
 	//------------------
 	// USER_CHECK
@@ -146,10 +231,10 @@ func userCreateDb(pInput *authUserCreateInput,
 
 //-------------------------------------------------------------
 // USER_GET__PIPELINE
-func userGetDb(pInput *authUserGetInput,
+func userGetDb(pInput *userGetInput,
 	pAuthenticatedBool bool,
 	pCtx context.Context,
-	pRuntime *runtime.Runtime) (*authUserGetOutput, error) {
+	pRuntime *runtime.Runtime) (*userGetOutput, error) {
 
 	//------------------
 	// VALIDATE
@@ -165,9 +250,9 @@ func userGetDb(pInput *authUserGetInput,
 		return nil, err
 	}
 
-	output := &authUserGetOutput{}
+	output := &userGetOutput{}
 	if pAuthenticatedBool {
-		output = &authUserGetOutput{
+		output = &userGetOutput{
 			UserNameStr: user.UserNameStr,
 			BioStr:      user.BioStr,
 		}
@@ -179,7 +264,7 @@ func userGetDb(pInput *authUserGetInput,
 }
 
 //-------------------------------------------------------------
-func userUpdateDb(pInput *authUserUpdateInput,
+func userUpdateDb(pInput *userUpdateInput,
 	pCtx context.Context,
 	pRuntime *runtime.Runtime) error {
